renamed components to uniform convention, moved images to img file, updaated imports
