+++ src/components/auth/login.test.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { BrowserRouter as Router } from "react-router-dom";
 import { render, fireEvent } from "@testing-library/react";
-import {getAllByText} from "@testing-library/dom";
+import { getAllByText } from "@testing-library/dom";
 import { getAllByTestId, getByTestId } from "../../utils/test-utils";
 
 // component to be tested.
-import Login from "./login";
+import Login from "./Login";
 
 // redux testing
 import { Provider } from "react-redux";
