+++ src/components/dashboard/navigation.test.js
@@ -4,7 +4,7 @@ import { BrowserRouter as Router } from "react-router-dom";
 import { getAllByTestId } from "../../utils/test-utils.js";
 
 // component to be tested.
-import Navigation from "./navigation.js";
+import Navigation from "./Navigation.js";
 
 // redux testing
 import { Provider } from "react-redux";
